refactor(console): refactor security page structure
refactor security page structure
