+++ packages/console/src/hooks/use-console-routes/routes/security.tsx
@@ -1,5 +1,7 @@
-import { type RouteObject } from 'react-router-dom';
+import { Navigate, type RouteObject } from 'react-router-dom';
 import { safeLazy } from 'react-safe-lazy';
+
+import { SecurityTabs } from '@/pages/Security/types';
 
 const Security = safeLazy(async () => import('@/pages/Security'));
 const CaptchaDetails = safeLazy(async () => import('@/pages/CaptchaDetails'));
@@ -7,10 +9,14 @@ const CaptchaDetails = safeLazy(async () => import('@/pages/CaptchaDetails'));
 export const security: RouteObject = {
   path: 'security',
   children: [
-    { index: true, element: <Security /> },
-    { path: 'guide/:guideId', element: <Security /> },
+    { index: true, element: <Navigate replace to={SecurityTabs.Captcha} /> },
+    { path: SecurityTabs.Captcha, element: <Security tab={SecurityTabs.Captcha} /> },
     {
-      path: 'captcha',
+      path: `${SecurityTabs.Captcha}/guide/:guideId`,
+      element: <Security tab={SecurityTabs.Captcha} />,
+    },
+    {
+      path: `${SecurityTabs.Captcha}/details`,
       element: <CaptchaDetails />,
     },
   ],
