+++ packages/console/src/pages/Security/Captcha/CaptchaCard/index.module.scss
