+++ packages/console/src/pages/Security/Captcha/CaptchaFormFields/index.tsx
