+++ packages/console/src/pages/Security/Captcha/CaptchaLogo/index.module.scss
