+++ packages/console/src/pages/Security/Captcha/CaptchaLogo/index.tsx
