+++ packages/console/src/pages/Security/Captcha/CreateCaptchaForm/ProviderRadioGroup/ProviderRadio/index.module.scss
