+++ packages/console/src/pages/Security/Captcha/CreateCaptchaForm/index.module.scss
