+++ packages/console/src/pages/Security/Captcha/CreateCaptchaForm/index.tsx
@@ -39,7 +39,7 @@ function CreateCaptchaForm({ onClose, isOpen: isFormOpen }: Props) {
             type="primary"
             disabled={!provider}
             onClick={() => {
-              navigate(`/security/guide/${provider}`);
+              navigate(`guide/${provider}`);
             }}
           />
         }
