+++ packages/console/src/pages/Security/Captcha/EnableCaptcha/index.module.scss
