+++ packages/console/src/pages/Security/Captcha/Guide/index.module.scss
