+++ packages/console/src/pages/Security/Captcha/index.module.scss
@@ -1,0 +1,12 @@
+@use '@/scss/underscore' as _;
+
+.content {
+  display: flex;
+  flex-direction: column;
+}
+
+.description {
+  font: var(--font-body-2);
+  color: var(--color-text-secondary);
+  margin-bottom: _.unit(3);
+}
