+++ packages/console/src/pages/Security/Captcha/types.tsx
