+++ packages/console/src/pages/Security/types.ts
@@ -1,0 +1,4 @@
+export enum SecurityTabs {
+  Captcha = 'captcha',
+  PasswordPolicy = 'password-policy',
+}
